requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
elasticsearch==8.15.1
sentence-transformers==2.6.1
pypdf==4.2.0
requests==2.32.3
ollama==0.3.1
python-dotenv==1.0.1
streamlit==1.37.0
pytest==8.3.2
google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
python-dotenv

elasticsearch>=8.13.0
requests



pytest>=8.0.0
